third_party/blink/web_tests/platform/generic/fast/forms/input-multiple-expected.txt

This tests that you can get and set the multiple attribute as a property in JavaScript.
 
Test 1 passed.
Test 2 passed.
Test 3 passed.
Test 4 passed.
Test 5 passed.
Test 6 passed.
